Converted new bug and new project forms to Formik and Yup
