+++ src/components/Account/AccountStatsCard.js
@@ -23,7 +23,7 @@ function AccountStatsCard({
 }) {
 	//--- JSX ---
 	return (
-		<Card className='card-corner-color-primary rounded hover-parent background-secondary full-width text-center'>
+		<Card className='rounded hover-parent full-width text-center'>
 			<Card.Body>
 				<Card.Title as='h2'>Statistics</Card.Title>
 				<div className='divider'></div>
