+++ src/components/Bugs/index.js
@@ -4,4 +4,4 @@
 export { default as AllBugs } from './AllBugs';
 export { default as BugCard } from './BugCard';
 export { default as Bug } from './Bug';
-export { default as NewBugForm } from './NewBugForm';
+export { default as ValidatedBugForm } from './ValidatedBugForm';
