+++ src/components/Projects/AllProjectsPage.js
@@ -27,22 +27,25 @@ function AllProjectsPage({ userProjects, currentUser }) {
 	return (
 		<Container fluid>
 			{currentUser !== null ? (
-				<Row id='project-cards-container'>
-					<Col className='col' lg={4} md={6} sm={12} xs={12}>
-						<Link to={`/projects/new`}>
-							<Card className='card new-project-card project-card rounded hover-parent background-primary card-corner-color-secondary'>
-								<Card.Body className='text-center'>
-									<Card.Title as='h1'>Create New Project</Card.Title>
-								</Card.Body>
-							</Card>
-						</Link>
-					</Col>
-					{userProjects !== undefined
-						? userProjects.map((project) => (
-								<ProjectCard key={project._id} project={project} />
-						  ))
-						: ''}
-				</Row>
+				<>
+					<h1 className='text-center'>My Projects</h1>
+					<Row id='project-cards-container'>
+						<Col className='col' lg={4} md={6} sm={12} xs={12}>
+							<Link to={`/projects/new`}>
+								<Card className='card new-project-card project-card rounded hover-parent background-primary card-corner-color-secondary'>
+									<Card.Body className='text-center'>
+										<Card.Title as='h1'>Create New Project</Card.Title>
+									</Card.Body>
+								</Card>
+							</Link>
+						</Col>
+						{userProjects !== undefined
+							? userProjects.map((project) => (
+									<ProjectCard key={project._id} project={project} />
+							  ))
+							: ''}
+					</Row>
+				</>
 			) : (
 				<div>
 					<p>
