+++ src/components/Projects/index.js
@@ -4,4 +4,4 @@
 export { default as AllProjectsPage } from './AllProjectsPage';
 export { default as ProjectCard } from './ProjectCard';
 export { default as ProjectPage } from './ProjectPage';
-export { default as NewProjectForm } from './NewProjectForm';
+export { default as ValidatedProjectForm } from './ValidatedProjectForm';
